Restructured and extended for individual use
* Module base for code that only depends on external projects
* Module common for code that depends on other LiberTEM common packages, but no core packages.
* Module udf for code that depends on LiberTEM core facilities

The intention is to allow code re-use without having to install all of LiberTEM, but also with
a reduced number of PyPI packages. Dependencies on LiberTEM common and core will be optional
features.

Still TODO:
* Modify setup.py accordingly
* Update documentation
