+++ src/libertem_blobfinder/__init__.py
@@ -1,22 +1,3 @@
-from .patterns import (
-    MatchPattern, Circular, RadialGradient, BackgroundSubtraction, UserTemplate,
-    RadialGradientBackgroundSubtraction
-)
-from .refinement import (
-    RefinementMixin, FastmatchMixin, AffineMixin, run_refine
-)
-from .correlation import (
-    get_peaks, CorrelationUDF, FastCorrelationUDF, FullFrameCorrelationUDF,
-    SparseCorrelationUDF, run_fastcorrelation, run_blobfinder
-)
-from .utils import feature_vector, visualize_frame, paint_markers
 from .__version__ import __version__
 
-__all__ = [
-    'MatchPattern', 'Circular', 'RadialGradient', 'BackgroundSubtraction', 'UserTemplate',
-    'RadialGradientBackgroundSubtraction',
-    'RefinementMixin', 'FastmatchMixin', 'AffineMixin', 'run_refine',
-    'get_peaks', 'CorrelationUDF', 'FastCorrelationUDF', 'FullFrameCorrelationUDF',
-    'SparseCorrelationUDF', 'run_fastcorrelation', 'run_blobfinder',
-    'feature_vector', 'visualize_frame', 'paint_markers', '__version__',
-]
+__all__ = ['__version__']
