+++ src/libertem_blobfinder/base/__init__.py
@@ -1,0 +1,2 @@
+from . import correlation
+__all__ = ['correlation']
