+++ src/libertem_blobfinder/common/__init__.py
@@ -1,0 +1,3 @@
+from . import correlation, patterns, utils
+
+__all__ = ['correlation', 'patterns', 'utils']
