+++ src/libertem_blobfinder/common/patterns.py
