+++ src/libertem_blobfinder/udf/__init__.py
@@ -1,0 +1,4 @@
+from . import correlation
+from . import refinement
+
+__all__ = ['correlation', 'refinement']
